+++ calendar-bundle/src/Resources/contao/classes/Events.php
@@ -255,12 +255,31 @@ abstract class Events extends Module
 
 			if (\is_array($arrRange) && isset($arrRange['unit']) && isset($arrRange['value']))
 			{
-				$strKey = 'cal_' . $arrRange['unit'];
-				$recurring = sprintf($GLOBALS['TL_LANG']['MSC'][$strKey], $arrRange['value']);
+				if ($arrRange['value'] == 1)
+				{
+					$repeat = $GLOBALS['TL_LANG']['MSC']['cal_single_'.$arrRange['unit']];
+				}
+				else
+				{
+					$repeat = sprintf($GLOBALS['TL_LANG']['MSC']['cal_multiple_'.$arrRange['unit']], $arrRange['value']);
+				}
 
 				if ($objEvents->recurrences > 0)
 				{
-					$until = sprintf($GLOBALS['TL_LANG']['MSC']['cal_until'], \Date::parse($objPage->dateFormat, $objEvents->repeatEnd));
+					$until = ' ' . sprintf($GLOBALS['TL_LANG']['MSC']['cal_until'], \Date::parse($objPage->dateFormat, $objEvents->repeatEnd));
+				}
+
+				if ($objEvents->recurrences > 0 && $intEnd < time())
+				{
+					$recurring = sprintf($GLOBALS['TL_LANG']['MSC']['cal_repeat_ended'], $repeat, $until);
+				}
+				elseif ($strTime)
+				{
+					$recurring = sprintf($GLOBALS['TL_LANG']['MSC']['cal_repeat'], $repeat, $until, date('Y-m-d\TH:i:sP', $intStart), \Date::parse($objPage->dateFormat, $intStart).', '.$strTime);
+				}
+				else
+				{
+					$recurring = sprintf($GLOBALS['TL_LANG']['MSC']['cal_repeat'], $repeat, $until, date('Y-m-d', $intStart), \Date::parse($objPage->dateFormat, $intStart));
 				}
 			}
 		}