+++ core-bundle/src/Resources/contao/library/Contao/InsertTags.php
@@ -882,6 +882,17 @@ class InsertTags extends Controller
 						$height = null;
 					}
 
+					// Use the alternative text from the image meta data if none is given
+					if (!$alt && ($objFile = \FilesModel::findByPath($strFile)))
+					{
+						$arrMeta = \Frontend::getMetaData($objFile->meta, $objPage->language);
+
+						if (isset($arrMeta['alt']))
+						{
+							$alt = $arrMeta['alt'];
+						}
+					}
+
 					// Generate the thumbnail image
 					try
 					{