+++ core-bundle/tests/Command/SymlinksCommandTest.php
@@ -14,16 +14,36 @@ namespace Contao\CoreBundle\Tests\Command;
 
 use Contao\CoreBundle\Command\SymlinksCommand;
 use Contao\CoreBundle\Config\ResourceFinder;
-use Contao\CoreBundle\ContaoCoreBundle;
 use Contao\CoreBundle\Tests\TestCase;
 use Symfony\Component\Console\Tester\CommandTester;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\Filesystem\Filesystem;
 use Symfony\Component\Lock\Factory;
 use Symfony\Component\Lock\Store\FlockStore;
 
 class SymlinksCommandTest extends TestCase
 {
+    /**
+     * @var SymlinksCommand
+     */
+    private $command;
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function setUp(): void
+    {
+        parent::setUp();
+
+        $this->command = new SymlinksCommand(
+            $this->getFixturesDir(),
+            'files',
+            $this->getFixturesDir().'/var/logs',
+            new ResourceFinder($this->getFixturesDir().'/vendor/contao/test-bundle/Resources/contao'),
+            $this->createMock(EventDispatcherInterface::class)
+        );
+    }
+
     /**
      * {@inheritdoc}
      */
@@ -47,21 +67,13 @@ class SymlinksCommandTest extends TestCase
 
         $finder = new ResourceFinder($this->getFixturesDir().'/vendor/contao/test-bundle/Resources/contao');
 
-        $container = $this->mockContainerWithBundles();
+        $container = $this->mockContainer($this->getFixturesDir());
         $container->setParameter('kernel.logs_dir', $this->getFixturesDir().'/var/logs');
         $container->set('contao.resource_finder', $finder);
 
-        $command = new SymlinksCommand(
-            $this->getFixturesDir(),
-            'files',
-            $this->getFixturesDir().'/var/logs',
-            new ResourceFinder($this->getFixturesDir().'/vendor/contao/test-bundle/Resources/contao'),
-            $container->getParameter('kernel.bundles')
-        );
+        $this->command->setContainer($container);
 
-        $command->setContainer($container);
-
-        $tester = new CommandTester($command);
+        $tester = new CommandTester($this->command);
         $code = $tester->execute([]);
         $display = $tester->getDisplay();
 
@@ -82,8 +94,6 @@ class SymlinksCommandTest extends TestCase
         $this->assertContains(' system/themes ', $display);
         $this->assertContains(' system/logs ', $display);
         $this->assertContains(' var/logs ', $display);
-        $this->assertContains(' system/config/tcpdf.php ', $display);
-        $this->assertContains('core-bundle/src/Resources/contao/config/tcpdf.php ', $display);
     }
 
     public function testIsLockedWhileRunning(): void
@@ -95,22 +105,13 @@ class SymlinksCommandTest extends TestCase
         }
 
         $factory = new Factory(new FlockStore($tmpDir));
-        $container = $this->mockContainerWithBundles();
 
         $lock = $factory->createLock('contao:symlinks');
         $lock->acquire();
 
-        $command = new SymlinksCommand(
-            $this->getFixturesDir(),
-            'files',
-            $this->getFixturesDir().'/var/logs',
-            new ResourceFinder($this->getFixturesDir().'/vendor/contao/test-bundle/Resources/contao'),
-            $container->getParameter('kernel.bundles')
-        );
+        $this->command->setContainer($this->mockContainer($this->getFixturesDir()));
 
-        $command->setContainer($this->mockContainer($this->getFixturesDir()));
-
-        $tester = new CommandTester($command);
+        $tester = new CommandTester($this->command);
         $code = $tester->execute([]);
 
         $this->assertSame(1, $code);
@@ -121,35 +122,17 @@ class SymlinksCommandTest extends TestCase
 
     public function testConvertsAbsolutePathsToRelativePaths(): void
     {
-        $container = $this->mockContainerWithBundles();
-
-        $command = new SymlinksCommand(
-            $this->getFixturesDir(),
-            'files',
-            $this->getFixturesDir().'/var/logs',
-            new ResourceFinder($this->getFixturesDir().'/vendor/contao/test-bundle/Resources/contao'),
-            $container->getParameter('kernel.bundles')
-        );
-
         // Use \ as directory separator in $rootDir
         $rootDir = new \ReflectionProperty(SymlinksCommand::class, 'rootDir');
         $rootDir->setAccessible(true);
-        $rootDir->setValue($command, strtr($this->getFixturesDir(), '/', '\\'));
+        $rootDir->setValue($this->command, strtr($this->getFixturesDir(), '/', '\\'));
 
         // Use / as directory separator in $path
         $method = new \ReflectionMethod(SymlinksCommand::class, 'getRelativePath');
         $method->setAccessible(true);
-        $relativePath = $method->invoke($command, $this->getFixturesDir().'/var/logs');
+        $relativePath = $method->invoke($this->command, $this->getFixturesDir().'/var/logs');
 
         // The path should be normalized and shortened
         $this->assertSame('var/logs', $relativePath);
-    }
-
-    private function mockContainerWithBundles(): ContainerBuilder
-    {
-        $container = $this->mockContainer($this->getFixturesDir());
-        $container->setParameter('kernel.bundles', ['ContaoCoreBundle' => ContaoCoreBundle::class]);
-
-        return $container;
     }
 }