+++ core-bundle/tests/Functional/app/AppKernel.php
@@ -31,7 +31,7 @@ use Symfony\Component\HttpKernel\Kernel;
 
 class AppKernel extends Kernel
 {
-    public function registerBundles()
+    public function registerBundles(): array
     {
         return [
             new FrameworkBundle(),
@@ -49,22 +49,22 @@ class AppKernel extends Kernel
         ];
     }
 
-    public function getProjectDir()
+    public function getProjectDir(): string
     {
         return \dirname(__DIR__);
     }
 
-    public function getRootDir()
+    public function getRootDir(): string
     {
         return __DIR__;
     }
 
-    public function getCacheDir()
+    public function getCacheDir(): string
     {
         return \dirname(__DIR__).'/var/cache/'.$this->environment;
     }
 
-    public function getLogDir()
+    public function getLogDir(): string
     {
         return \dirname(__DIR__).'/var/logs';
     }